Finish refactor to abstract tracked job
